FIX: Allow the use of stable tags
Pinning to a specific branch on a reusable module is an antipattern as
It is very flexible - projects will be blocked from installing stable
tags.

If there is code in a future branch that you need, then you should pin
to a not-yet-existing branch, and lower the min-stability of your parent
project to development.

Better yet, the modules with necessary code should have a stable tag
created.