Recommend this module only when using behat
When you're not using behat for end-to-end testing, I would recommend that we avoid the use of this recipe in favour of simply inlining the two other relevant requirements. It has a complex dependency chain that can conflicts with other packages (eg https://github.com/silverstripe/silverstripe-event-dispatcher/issues/8) and is generally only worth the trouble if you're actually using behat for testing.